Revert "add more BerH Vorabcheck tests, move machine utils to util file, add testcases property to Flow, simplify testcases interface, add flow coverage test"
This reverts commit ac9407964bcbf734eca37a76374ec376cb530dd1.
