+++ app/domains/__test__/TestCases.ts
@@ -1,3 +1,4 @@
+import type { Config } from "~/services/flow/server/types";
 import type { UserData } from "../userData";
 
 // Old flow tests: forward & backward using full user data
@@ -6,10 +7,13 @@ export type TestCases<T extends UserData> = Readonly<
 >;
 
 // New flow tests: testing data submission with page schemas
-export type FlowTestCases = Record<
-  string,
-  Array<{
-    stepId: string;
-    userInput?: UserData;
-  }>
->;
+export type FlowTestCases = {
+  xstateConfig: Config;
+  testcases: Record<
+    string,
+    Array<{
+      stepId: string;
+      userInput?: UserData;
+    }>
+  >;
+};
