+++ app/domains/__test__/flows.test.ts
@@ -1,3 +1,5 @@
+import { toDirectedGraph, type DirectedGraphNode } from "@xstate/graph";
+import { pathToStateValue } from "xstate";
 import { testCasesBeratungshilfeFormular } from "~/domains/beratungshilfe/formular/__test__/testCasesBeratungshilfeFormular";
 import { testCasesFluggastrechteFormular } from "~/domains/fluggastrechte/formular/__test__/testCasesFluggastrechteFormular";
 import { testCasesFluggastrechteVorabcheck } from "~/domains/fluggastrechte/vorabcheck/__test__/testCasesFluggastrechteVorabcheck";
@@ -7,12 +9,12 @@ import {
   testCasesProzesskostenhilfeSubmitOnly,
 } from "~/domains/prozesskostenhilfe/formular/__test__/testcases";
 import { type UserData } from "~/domains/userData";
-import { allStepsFromMachine } from "~/services/flow/machineUtils";
 import { nextStepId } from "~/services/flow/server/buildFlowController";
 import type {
   FlowStateMachine,
   NavigationEvent,
 } from "~/services/flow/server/types";
+import { stateValueToStepIds } from "~/services/flow/stepIdConverter";
 
 function getEnabledSteps({
   machine,
@@ -36,6 +38,21 @@ function getEnabledSteps({
     return destination;
   });
   return [initialStep, ...reachableSteps];
+}
+
+function statePathsFromMachine(children: DirectedGraphNode[]): string[][] {
+  return children.flatMap((child) =>
+    child.children.length > 0
+      ? statePathsFromMachine(child.children)
+      : [child.stateNode.path],
+  );
+}
+
+function allStepsFromMachine(machine: FlowStateMachine) {
+  const machineState = statePathsFromMachine(toDirectedGraph(machine).children);
+  return machineState.map(
+    (statePath) => stateValueToStepIds(pathToStateValue(statePath))[0],
+  );
 }
 
 /*
