+++ app/domains/beratungshilfe/vorabcheck/index.ts
@@ -3,7 +3,6 @@ import { sendCustomAnalyticsEvent } from "~/services/analytics/customEvent";
 import { getVerfuegbaresEinkommenFreibetrag } from "./freibetrag";
 import { type BeratungshilfeVorabcheckUserData } from "./userData";
 import { beratungshilfeVorabcheckXstateConfig } from "./xstateConfig";
-import { beratungshilfeVorabcheckTestCases } from "~/domains/beratungshilfe/vorabcheck/__test__/testcasesWithUserInputs";
 
 export const beratungshilfeVorabcheck = {
   flowType: "vorabCheck",
@@ -12,7 +11,6 @@ export const beratungshilfeVorabcheck = {
       getVerfuegbaresEinkommenFreibetrag(context).toString(),
   }),
   config: beratungshilfeVorabcheckXstateConfig,
-  testcases: beratungshilfeVorabcheckTestCases,
   guards: {},
   asyncFlowActions: {
     "/bereich": (request, userData: BeratungshilfeVorabcheckUserData) =>
